lab-recap: Fix lab exercises
    * Add `.gitignore` files
    * Faile compile-time errors
    * Improve coding style
    * Remove year from the block comments at the beginning of C files
    * Rework exercise 1. The previous exercise was broken.

+++ labs/recap/skel/04-functions-hijacking/test.c
@@ -1,3 +1,10 @@
+/**
+ * SO
+ * Recap
+ * Task #4
+ *
+ * Driver program to test function hijacking
+ */
 #include <stdio.h>
 #include <stdlib.h>
 #include <pthread.h>
@@ -6,21 +13,21 @@
 
 static void *thread_code(void *arg)
 {
-    printf("Hello from the user thread code\n");
-    return NULL;
+	printf("Hello from the user thread code\n");
+	return NULL;
 }
 
 int main()
 {
-    pthread_t threads[N];
-    int i;
+	pthread_t threads[N];
+	size_t i;
 
-    for (i = 0 ; i < N ; i++)
-        pthread_create(&threads[i], NULL, thread_code, (void *) i);
+	for (i = 0 ; i < N ; i++)
+		pthread_create(&threads[i], NULL, thread_code, (void *)i);
 
-    for (i = 0 ; i < N ; i++)
-        pthread_join(threads[i], NULL);
+	for (i = 0 ; i < N ; i++)
+		pthread_join(threads[i], NULL);
 
-    return 0;
+	return 0;
 }
